fix: cross-platform compilation for Linux (Swift 6.0)
- Upgrade to swift-tools-version 6.0, Docker uses swift:6.0-noble
- Build SQLite from source with SQLITE_ENABLE_SNAPSHOT for GRDB
- Platform guards: PTYProcess, PTYMultiplexer, HookServer, ControlServer
  wrapped with #if canImport(Darwin)
- Replace CGFloat with Double in cross-platform protocols/models
- Fix Hummingbird 2.0 API: ByteBuffer, request.authority, NIOCore imports
- CLI target moved to macOS-only (URLSession incomplete on Linux)
- Tests: guard AppKit imports and macOS-only test suites
- Add FoundationNetworking import for Linux CLI
- Cross-platform timer in PerformanceBenchmarks

+++ Sources/Core/Rendering/MetalRenderer.swift
@@ -61,8 +61,8 @@ public final class MetalRenderer: NSObject, TerminalRenderable, MTKViewDelegate 
         pendingViewports = viewports
     }
 
-    public func updateGlyphAtlas(fontFamily: String, fontSize: CGFloat) {
-        glyphAtlas.build(fontFamily: fontFamily, fontSize: fontSize)
+    public func updateGlyphAtlas(fontFamily: String, fontSize: Double) {
+        glyphAtlas.build(fontFamily: fontFamily, fontSize: CGFloat(fontSize))
     }
 
     public func invalidate() {
@@ -393,7 +393,7 @@ public final class MetalRenderer: NSObject, TerminalRenderable, MTKViewDelegate 
 public final class NullRenderer: TerminalRenderable {
     public init() {}
     public func render(viewports: [TerminalViewportData]) {}
-    public func updateGlyphAtlas(fontFamily: String, fontSize: CGFloat) {}
+    public func updateGlyphAtlas(fontFamily: String, fontSize: Double) {}
     public func invalidate() {}
 }
